refactored transformation and aggregation
